add Sepolia testnet to production blockchain list
