+++ connectors/chains.ts
@@ -439,7 +439,7 @@ export type Blockchain = keyof typeof RPC;
 
 export const RPCList = Object.values(RPC)
   // filter out testnets in prod, except for Goerli
-  .filter((chain) => !isProdEnv || !chain.testnet || chain.chainId === goerli.id);
+  .filter((chain) => !isProdEnv || !chain.testnet || chain.chainId === goerli.id || chain.chainId === sepolia.id);
 
 export function getChainShortname(chainId: string | number): string {
   const parsedChainId = parseInt(chainId.toString());
