+++ package.json
@@ -125,7 +125,7 @@
     "@lens-protocol/metadata": "^1.1.5",
     "@lens-protocol/react-web": "^2.0.0-alpha.27",
     "@lens-protocol/wagmi": "^3.0.0-alpha.24",
-    "@lit-protocol/constants": "^2.2.61",
+    "@lit-protocol/constants": "^2.2.63",
     "@lit-protocol/lit-node-client": "^2.2.61",
     "@metamask/onboarding": "^1.0.1",
     "@mui/icons-material": "^5.15.7",
